fix: Issue # 49
While calling getChildren with offset and limitCount, in condition when the offset is close to the total items in a folder, such that the request limit becomes less than the actual remaining limit out of toal items, it throws an assertion at LOC 593.

The System.assertEquals(totalCount, children.size()); needs to be changed to something to handle if the offset it much larger more than the actual remaining limit.

The issue is fixed with a condition added to totalCount within an assertion.